ERT-1001 Restructured config nodes
