+++ devel/python/python/ert/enkf/config/__init__.py
@@ -1,0 +1,12 @@
+from .custom_kw_config import CustomKWConfig
+from .field_config import FieldConfig
+from .gen_data_config import GenDataConfig
+from .gen_kw_config import GenKwConfig
+from .summary_config import SummaryConfig
+from .enkf_config_node import EnkfConfigNode
+
+__all__ = ["FieldConfig",
+           "CustomKWConfig",
+           "GenKwConfig",
+           "GenDataConfig",
+           "EnkfConfigNode",]
