+++ devel/python/python/ert/enkf/config/gen_data_config.py
