+++ devel/python/python/ert/enkf/data/__init__.py
@@ -1,24 +1,11 @@
-from .gen_data_config import GenDataConfig
-from .gen_kw_config import GenKwConfig
-from .field_config import FieldConfig
-from .summary_config import SummaryConfig
-from .custom_kw_config import CustomKWConfig
-from .field import Field
-
-from .enkf_config_node import EnkfConfigNode
 from .custom_kw import CustomKW
-from .gen_kw import GenKw
+from .field import Field
 from .gen_data import GenData
+from .gen_kw import GenKw
 from .enkf_node import EnkfNode
 
-__all__ = ["GenDataConfig",
-           "GenKwConfig",
-           "FieldConfig",
-           "SummaryConfig",
-           "CustomKWConfig",
-           "Field",
-           "EnkfConfigNode",
-           "EnkfNode",
+__all__ = ["Field",
            "CustomKW",
            "GenKw",
-           "GenData"]
+           "GenData",
+           "EnkfNode",]
