+++ devel/python/python/ert/enkf/data/field.py
@@ -17,7 +17,7 @@ import sys
 
 from ert.cwrap import BaseCClass, CWrapper
 from ert.enkf import ENKF_LIB
-from ert.enkf.data import FieldConfig
+from ert.enkf.config import FieldConfig
 from ert.enkf.enums import EnkfStateType
