+++ devel/python/tests/ert/enkf/data/test_custom_kw_config.py
@@ -1,4 +1,4 @@
-from ert.enkf.data.custom_kw_config import CustomKWConfig
+from ert.enkf.config.custom_kw_config import CustomKWConfig
 from ert.test.extended_testcase import ExtendedTestCase
 from ert.test.test_area import TestAreaContext
 from ert.util import StringList
